+++ example/hanging_beam/hanging_beam_continuum_2d.jl
@@ -1,12 +1,15 @@
 #=
   @ author: bcynuaa <bcynuaa@163.com> | callm1101 <Calm.Liu@outlook.com> | vox-1 <xur2006@163.com>
-  @ date: 2024/07/14 16:47:45
+  @ date: 2024/07/19 20:26:57
   @ license: MIT
   @ description:
  =#
 
 using EtherSPH
 using ProgressBars
+using PyCall
+
+# see `SPH/Cart/LargeDeformationContinuumModel.jl`
 
 const dim = 2
 const beam_width = 0.02
@@ -14,18 +17,10 @@ const beam_length = 10 * beam_width
 const beam_inside_length = 3 * beam_width
 const dr = beam_width / 20
 const gap = dr
-const h = 1.6 * dr
+const h = 1.5 * dr
 const wall_width = 3 * dr
 
-const kernel = WendlandC4{dim}(h)
-
-@inline function W(r::Float64)::Float64
-    return kernelValue(r, kernel)
-end
-
-@inline function DW(r::Float64)::Float64
-    return kernelGradient(r, kernel)
-end
+const kernel = CubicSpline{dim}(h)
 
 const young_modulus = 2e6
 const poisson_ratio = 0.3975
@@ -65,7 +60,7 @@ const fl = f(l)
     end
 end
 
-const dt = 0.2 * h / c_0
+const dt = 0.1 * h / c_0
 const total_time = 1.0
 const output_dt = 1000 * dt
 @info "total steps: $(round(Int, total_time / dt))"
@@ -98,71 +93,30 @@ const output_dt = 1000 * dt
     corrected_piola_kirchhoff_1st_stress_mat_::Matrix2D = Matrix0(dim) # Pᶜ
 end
 
-# after create neighbour index list ...
-
-@inline function initialElasticParticles!(particle_system::ParticleSystem{Dimension2, Particle}, p::Particle)::Nothing
-    p.rho_0_ = p.rho_
-    p.volume_0_ = p.mass_ / p.rho_
-    p.initial_neighbour_index_list_ = deepcopy(p.neighbour_index_list_)
-    @inbounds @simd for i in eachindex(p.initial_neighbour_index_list_)
-        q_index = p.initial_neighbour_index_list_[i]
-        q = particle_system[q_index]
-        rpq = p.neighbour_position_list_[i]
-        r = p.neighbour_distance_list_[i]
-        dw = DW(r)
-        r_inv = 1 / r
-        push!(p.kernel_gradient_0_vec_list_, dw * r_inv * rpq)
-        p.corrected_mat_ .+= -q.volume_0_ * dw * r_inv * dyad(rpq, rpq)
-    end
-    p.corrected_mat_ .= inv(p.corrected_mat_) # B⁰
-    @inbounds @simd for i in eachindex(p.initial_neighbour_index_list_)
-        push!(p.corrected_kernel_gradient_0_vec_list_, dot(p.kernel_gradient_0_vec_list_[i], p.corrected_mat_))
-    end
+@inline function initializeContinuum!(particle_system::ParticleSystem{dim, Particle}, p::Particle)::Nothing
+    EtherSPH.libLDCMInitializeContinuum!(particle_system, p; smooth_kernel = kernel)
     return nothing
 end
 
-@inline function calculateDeformationMat!(particle_system::ParticleSystem{Dimension2, Particle}, p::Particle)::Nothing
-    @inbounds @simd for i in eachindex(p.kernel_gradient_0_vec_list_)
-        q = particle_system[p.initial_neighbour_index_list_[i]]
-        p.deformation_gradient_mat_ .+=
-            q.volume_0_ * dt * dyad((q.v_vec_ .- p.v_vec_), p.corrected_kernel_gradient_0_vec_list_[i])
-    end
-    p.rho_ = p.rho_0_ / det(p.deformation_gradient_mat_)
-    p.green_lagrange_strain_mat_ .=
-        0.5 * (p.deformation_gradient_mat_' * p.deformation_gradient_mat_ .- ConstMatI{dim}())
-    p.piola_kirchhoff_2nd_stress_mat_ .=
-        lambda * trace(p.green_lagrange_strain_mat_) * ConstMatI{dim}() .+ 2 * mu * p.green_lagrange_strain_mat_
-    p.piola_kirchhoff_1st_stress_mat_ .= p.deformation_gradient_mat_ * p.piola_kirchhoff_2nd_stress_mat_
-    p.corrected_piola_kirchhoff_1st_stress_mat_ .= p.piola_kirchhoff_1st_stress_mat_ * p.corrected_mat_
+@inline function calculateDeformationGradient!(particle_system::ParticleSystem{dim, Particle}, p::Particle;)::Nothing
+    EtherSPH.libLDCMDeformationGradientEvolution!(particle_system, p; dt = dt)
+    EtherSPH.libLDCMGreenLagrangeStrain!(p)
+    EtherSPH.libLDCMLinearElasticityStress!(p; lambda = lambda, mu = mu)
+    EtherSPH.libLDCMPioalKirchhoffStress!(p)
     return nothing
 end
 
-@inline function momentum!(particle_system::ParticleSystem{Dimension2, Particle}, p::Particle)::Nothing
+@inline function momentum!(particle_system::ParticleSystem{dim, Particle}, p::Particle;)::Nothing
     if p.type_ == MOVABLE_MATERIAL_TAG
-        @inbounds @simd for i in eachindex(p.kernel_gradient_0_vec_list_)
-            q = particle_system[p.initial_neighbour_index_list_[i]]
-            rpq = p.x_vec_ .- q.x_vec_
-            r = norm(rpq)
-            v_dot_x = dot(p.v_vec_ .- q.v_vec_, rpq)
-            if v_dot_x > 0.0
-                artificial_stress = 0.0
-            else
-                mean_rho = 0.5 * (p.rho_ + q.rho_)
-                mean_c = 0.5 * (p.c_ + q.c_)
-                phi = h * v_dot_x / (r * r + 0.01 * h * h)
-                artificial_stress = (-artificial_alpha * mean_c * phi + artificial_beta * phi * phi) / mean_rho
-            end
-            p.dv_vec_ .+=
-                1 / p.mass_ *
-                p.volume_0_ *
-                q.volume_0_ *
-                dot(
-                    p.corrected_piola_kirchhoff_1st_stress_mat_ .+ q.corrected_piola_kirchhoff_1st_stress_mat_,
-                    p.kernel_gradient_0_vec_list_[i],
-                )
-            p.dv_vec_ .-= q.mass_ * artificial_stress * DW(r) / r * rpq
-        end
-        return nothing
+        EtherSPH.libLDCMContinuumMomentum!(particle_system, p)
+        EtherSPH.libLDCMContinuumArtificialStress!(
+            particle_system,
+            p;
+            smooth_kernel = kernel,
+            alpha = artificial_alpha,
+            beta = artificial_beta,
+            h = h,
+        )
     end
     return nothing
 end
@@ -170,7 +124,6 @@ end
 @inline function accelerateAndMove!(p::Particle)::Nothing
     if p.type_ == MOVABLE_MATERIAL_TAG
         EtherSPH.libAccelerateAndMove!(p; dt = dt)
-        return nothing
     end
     return nothing
 end
@@ -213,7 +166,6 @@ append!(system, particles)
 vtp_writer = VTPWriter()
 @inline getVelocity(p::Particle)::Vector2D = p.v_vec_
 addVector!(vtp_writer, "Velocity", getVelocity)
-
 @inline getDeformationMatDet(p::Particle)::Float64 = det(p.deformation_gradient_mat_)
 addScalar!(vtp_writer, "DeformationMatDet", getDeformationMatDet)
 @inline getDeformationX(p::Particle)::Vector2D =
@@ -222,23 +174,24 @@ addVector!(vtp_writer, "DeformationX", getDeformationX)
 @inline getDeformationY(p::Particle)::Vector2D =
     Vector2D(p.deformation_gradient_mat_[3], p.deformation_gradient_mat_[4])
 addVector!(vtp_writer, "DeformationY", getDeformationY)
-
 @inline getVonMisesStress(p::Particle)::Float64 =
     sqrt(1.5 * ddot(p.piola_kirchhoff_2nd_stress_mat_, p.green_lagrange_strain_mat_))
 addScalar!(vtp_writer, "VonMisesStress", getVonMisesStress)
 
+@info "$vtp_writer"
+
 vtp_writer.step_digit_ = 4
-vtp_writer.file_name_ = "deformation_2d_"
-vtp_writer.output_path_ = "demo/results/deformation_2d"
+vtp_writer.file_name_ = "hanging_beam_continuum_2d_"
+vtp_writer.output_path_ = "example/results/hanging_beam/hanging_beam_continuum_2d"
 
 function main()::Nothing
     t = 0.0
     assurePathExist(vtp_writer)
     createNeighbourIndexList!(system)
-    applyReflection!(system, initialElasticParticles!)
+    applyReflection!(system, initializeContinuum!)
     saveVTP(vtp_writer, system, 0, t)
     for step in ProgressBar(1:round(Int, total_time / dt))
-        applyReflection!(system, calculateDeformationMat!)
+        applyReflection!(system, calculateDeformationGradient!)
         applyReflection!(system, momentum!)
         applySelfaction!(system, accelerateAndMove!)
         if step % round(Int, output_dt / dt) == 0
@@ -246,5 +199,14 @@ function main()::Nothing
         end
         t += dt
     end
+    return nothing
+end
+
+function post()::Nothing
+    PyCall.@pyinclude "example/hanging_beam/hanging_beam_continuum_2d.py"
+    HangingBeam2DPostProcess = py"HangingBeam2DPostProcess"
+    post_process = HangingBeam2DPostProcess(reference_gap = gap)
+    post_process.viewPlot()
+    post_process.referencePlot()
     return nothing
 end