+++ CHANGELOG.md
@@ -10,7 +10,154 @@ All notable changes to this project will be documented in this file. The format 
 
 ---
 
-## [Unreleased] - 2024-07-14
+## [0.1.0] - 2024-07-xx
+
+### Brief
+
+#### Call `ConstVec` or `ConstMat`
+
+In continuum mechanics, `ConstVec` or `ConstMat` is recommended to use for constant vectors or matrices. We need to use a vector or matrix with different dimensions in calculation.
+
+How it works? Let's see a simple example:
+
+```julia
+const MatrixI2D = MatrixI(2)
+const MatrixI3D = MatrixI(3)
+
+struct ConstMatI{Dimension} end
+@inline function ConstMatI{Dimension2}()::Matrix2D
+    return MatrixI2D
+end
+@inline function ConstMatI{Dimension3}()::Matrix3D
+    return MatrixI3D
+end
+```
+
+The struct `ConstMatI` is defined to call `MatrixI2D` or `MatrixI3D` according to the dimension option. We can use a constant matrix $\mathrm{I}$ like:
+
+```julia
+A = Matrix2D(0., 0., 0., 0.)
+A .= ConstMatI{2}() # without any cost
+```
+
+Supported types:
+
+- `ConstVec0{Dimension}`
+- `ConstVecX{Dimension}`
+- `ConstVecY{Dimension}`
+- `ConstVecZ{Dimension}`
+- `ConstMatI0Dimension}`
+- `ConstMatI{Dimension}`
+
+#### `TinyLinearAlgebra` for GC problems
+
+In our tests, `MArray` from `StaticArrays` suffers from GC problems. A small example is that:
+
+```julia
+x = Vector2D(1., 2.)
+A = Matrix2D(1., 2., 3., 4.)
+c = Vector2D(0., 0.)
+B = Matrix2D(0., 0., 0., 0.)
+
+c .= A * x # suffers from GC problems in for-loop
+c .= x' * A # suffers from GC problems in for-loop
+c .= dot(A, x) # recommended
+c .= dot(x, A) # recommended
+
+B .= x * x' # suffers from GC problems in for-loop
+B .= dyad(x, x) # recommended
+
+B .= A * A # recommended, the same as below ↓
+B .= A * A' # recommended
+B .= dot(A, A) # recommended
+B .= dot(A, A') # recommended
+```
+It seems that gc problems are caused by converting `Vector` to `Matrix` or `Matrix` to `Vector`. Always carefully check the GC problems in your code. If you find abnormal speed down in your code, it's recommended use `profview` to check the GC problems. Functions in `Math` are recommended to use `TinyLinearAlgebra` to avoid GC problems.
+
+#### `VTPWriter` for field data
+
+Sometimes, additional field data in one vtp file is needed. We introduce an optional argument `field_dict::Dict{String, Any} = Dict{String, Any}()` in `saveVTP` function to save field data in VTP file. Construct a `Dict` and parse it into `saveVTP` function to save field data in VTP file.
+
+#### `NeighbourSearch` and `Apply` for continuum mechanics
+
+In continuum mechanics, time-varied neighbour search is not needed. We should record all neighbour index first (ensure the index not changes in your particle system). Then, some function should be applied on particles in each time step.
+
+For the order of neighbour particle indexes contained by one particle matters, the particle action should be applied with the sequence given. `directlyUpdateNeighbours!` is introduced to update neighbour particles in continuum mechanics. `applyReflection!` is introduced to apply reflection particle action if related particle ids are recorded. The code is simply like:
+
+```julia
+@inline function directlyUpdateNeighbours!(
+    particle_system::ParticleSystem{Dimension, ParticleType},
+)::Nothing where {Dimension, ParticleType <: AbstractParticle{Dimension}}
+    Threads.@threads for i in eachindex(particle_system)
+        p = particle_system.particles_[i]
+        @simd for j in eachindex(p.neighbour_index_list_)
+            @inbounds q_index = p.neighbour_index_list_[j]
+            @inbounds p.neighbour_position_list_[j] .= p.x_vec_ .- particle_system[q_index]
+            @inbounds p.neighbour_distance_list_[j] = norm(p.neighbour_position_list_[j])
+        end
+    end
+    return nothing
+end
+
+@inline function applyReflection!(
+    particle_system::ParticleSystem{Dimension, ParticleType},
+    reflectionFunction!::Function;
+    parameters...,
+)::Nothing where {Dimension, ParticleType <: AbstractParticle{Dimension}}
+    Threads.@threads for i in eachindex(particle_system)
+        @inbounds reflectionFunction!(particle_system, particle_system[i]; parameters...)
+    end
+    return nothing
+end
+```
+
+#### Continuum Mechanics Model
+
+In `SPH/Cart/LargeDeformationContinuumMechanics.jl`, a large deformation continuum mechanics model is introduced. Any elastic material can now be calculated with the model. The model is based on the continuum mechanics theory and the SPH discretization. The calculation steps is briefly shown as below:
+
+- initialize the neighbour particles and kernel gradient.
+- calculate the deformation gradient and green-lagrange strain.
+- use constitutive relation (either elastic or plastic) to calculate stress.
+- calculate the acceleration of particles with its stress (or artificial stress should be added)
+
+Examples can be found [here: haning beam continuum](example/hanging_beam/hanging_beam_continuum_2d.jl). A [demo](demo/deformation_2d/deformation_2d.jl) when I was developing such model is also included.
+
+To achieve the model, `SPH/Library` folder adds:
+
+- `SPH/Library/ViscosityForce.jl` adds `libArtificialViscosityForce!` function to apply artificial stress to avoid particles being too close to each other.
+- `SPH/Library/Correction.jl` is added with corrected matrix calculated. In paper, such matrix is usually denoted as $\mathbf{B}$ or $\mathbf{L}^{-1}$.
+
+### Added
+
+- `Math/Vector.jl` adds `ConstVec` xxx.
+- `Math/Matrix.jl` adds `ConstMat` xxx.
+- `Math/TinyLinearAlgebra.jl` adds `dot` for `Matrix` and `Vector`.
+- `Post/VTPWriter.jl/saveVTP` function adds an optional argument .`field_dict::Dict{String, Any} = Dict{String, Any}()` to save field data in VTP file.
+- `Core/NeighbourSearch.jl` adds `directlyUpdateNeighbours!` for continuum mechanics.
+- `Core/Apply.jl` adds `applyReflection!` for users to apply reflection
+particle action if related particle ids are recorded.
+- `SPH/Cart/Cart.jl` is added which is just like a cart in shopping mall. Feel free to add model described in papers here!
+- `SPH/Cart/LargeDeformationContinuumMechanics.jl` adds a large deformation continuum mechanics model.
+- `SPH/Library/ViscosityForce.jl` adds `libArtificialViscosityForce!` function to apply artificial stress.
+- `SPH/Library/Correction.jl` adds corrected matrix calculation.
+- `demo/deformation_2d/deformation_2d.jl` is added to show how I tested and developed the large deformation continuum mechanics model.
+- `example/hanging_beam/hanging_beam_continuum_2d.jl` is added to show how to use the large deformation continuum mechanics model contained in `src/SPH/Cart/LargeDeformationContinuumMechanics.jl`.
+
+### Changed
+
+- `Math/Matrix.jl/dot` is changed to `Math/Matrix.jl/ddot` to avoid conflict.
+
+### Fixed
+
+- Nothing
+
+### Removed
+
+- Nothing
+
+---
+
+## [0.1.0] - 2024-07-14
 
 ### Brief